% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valuator.R
\name{calculate_lifetime_value}
\alias{calculate_lifetime_value}
\title{Calculate Lifetime Value - Simplified}
\usage{
calculate_lifetime_value(gc, r, d)
}
\arguments{
\item{gc}{an atomic vector representing the rank of the values.}

\item{r}{an atomic vector representing the value itself.}

\item{d}{an atomic vector representing the number of occurrences of that value.}
}
\value{
A character vector where each element is the concatenation of rank, value, and number of occurrences.
}
\description{
\code{concat_rank_value_occurrence} takes vectors representing rank, value, and number of occurrences and returns them in a concatenated string.
}
\examples{
# Let's make a comment
concat_rank_value_occurrence("value rank", "value", "number of value occurrences")
concat_rank_value_occurrence(1, "green", "15")
concat_rank_value_occurrence(5, 6, 7)
}

