% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valuator.R
\docType{package}
\name{foo}
\alias{foo}
\alias{foo-package}
\title{valuator: A package for valuing customers.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

