% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/valuator.R
\name{frequent_values}
\alias{frequent_values}
\title{Least or Most Frequent Values}
\usage{
frequent_values(x, n = 5L, type = "most")
}
\arguments{
\item{x}{an atomic vector.}

\item{n}{a single positive integer indicating the number of values to return.}

\item{type}{a character string specifying which method should be used.}
}
\value{
A character vector of length \code{n} that returns the least or most \code{n} occuring values.
}
\description{
\code{concat_rank_value_occurrence} takes a vector and returns the \code{n} least or most occuring values.
}
\examples{
x <- c(letters, letters[1:4])
frequent_values(x, n = 5L, type = "most")
}

